+++ function-trigger-deployer-rs/src/dynamodb_manager/tests/custom_container/dynamo_test.rs
