+++ function-trigger-deployer-rs/src/dynamodb_manager/tests/custom_container/mod.rs
@@ -1,0 +1,1 @@
+pub mod dynamo_test;
