--- function-trigger-deployer-rs/src/tests/custom_container/mod.rs
@@ -1,1 +1,0 @@
-pub mod dynamo_test;
